move decode into MessageDecoder
